# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Accelyo API",
    "description": "Plateforme de carte etudiante dematerialisee. Toutes les routes (hors /auth/login, /auth/refresh et /nfc/validate) exigent un Bearer JWT.",
    "version": "1.0.0",
    "contact": {
      "name": "Accelyo",
      "email": "support@accelyo.fr"
    }
  },
  "servers": [
    {
      "url": "https://api.accelyo.fr/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Developpement local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "readerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Reader-Api-Key",
        "description": "Cle API d un lecteur Elatec (route /nfc/validate)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cards/{id}/revoke": {
      "post": {
        "tags": [
          "Cards"
        ],
        "summary": "Revoquer une carte (motif obligatoire)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Revoquee"
          },
          "400": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}